question_1/pyproject.toml
[project]
name = "tymex"
version = "0.0.1"
description = ""
authors = [
    {name = "sam-val",email = "code@samval.dev"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "fastapi-pagination (>=0.13.2,<0.14.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=37.4.0,<38.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-env (>=1.1.5,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
]

[[tool.poetry.packages]]
include = "scripts"
 
[tool.poetry.scripts]
shell = "scripts.shell:main"
 
[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90

[tool.isort]
profile = "black"
